// generated/models/CustomerAttributes.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Metadata } from './Metadata';
export type CustomerAttributes = {
    account_id?: string;
    name?: string;
    email?: string;
    phone?: string;
    metadata?: Metadata;
};