// generated/models/CustomerResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Metadata } from './Metadata';
export type CustomerResponse = {
    id?: string;
    account_id?: string;
    email?: string;
    metadata?: Metadata;
    name?: string;
    payment_method_id?: string;
    phone?: string;
    type?: string;
    created_at?: string;
    updated_at?: string;
};